// types/firebase.ts
export interface Stats {
    month_calls: number
}
export interface GlobalStats {
    current_month_calls: number
    average: number
    median: number
    history: Record<string, Stats>
}

/* export class GlobalStats implements IGlobalStats {
    current_month_calls: number
    average: number
    median: number
    history: Record<string, Stats>

    constructor(current_month_calls: number = 0, average: number = 0, median: number = 0, history: Record<string, Stats> = {}) {
        this.current_month_calls = current_month_calls
        this.average = average
        this.median = median
        this.history = history
    }



}
 */